+++ src/vgr/stmt_log.py
@@ -5,11 +5,17 @@ Implementation of Log statements
 import logging
 from lark import Tree
 
+from .app_exceptions import VgrRuntimeError
 from .builtins import poly_format, bound_ops
-from .dd_config import LOG_LEVEL_PATH
+from .dd_config import VGR_PREFIX
 from .exec_context import ExecContext
+from .log_config import LOG_LEVEL_MAP
 
 _USER_LOGGER = logging.getLogger('vgr_user')
+
+_LOG_FILE_PATH = (VGR_PREFIX, 'log', 'file')
+_LOG_LEVEL_PATH = (VGR_PREFIX, 'log', 'level')
+_DEFAULT_LOG_LEVEL = "Info"
 
 _LOG_FUNCTION = {
     "Debug":    _USER_LOGGER.debug,
@@ -20,25 +26,13 @@ _LOG_FUNCTION = {
     "Critical": _USER_LOGGER.critical,
 }
 
-_LEVEL_MAP = {
-    "Debug":    logging.DEBUG,
-    "Info":     logging.INFO,
-    "Warn":     logging.WARNING,
-    "Warning":  logging.WARNING,
-    "Error":    logging.ERROR,
-    "Critical": logging.CRITICAL,
-    "Off":      logging.CRITICAL + 1,
-}
-
 # Doc combined with execute_log
 def execute_log_setlevel(ctx: ExecContext, statement: Tree) -> None:
-    log_level = statement.children[0].data.title()
-    level = _LEVEL_MAP.get(log_level)
+    level_name = statement.children[0].data.title()
+    level = LOG_LEVEL_MAP.get(level_name)
     if level is None:
-        raise ValueError(f'Unsupported log level {log_level!r}')
-    _USER_LOGGER.setLevel(level)
-    ctx.set_var(log_level, *LOG_LEVEL_PATH)
-    ctx.print_verbose('Log Level set to', log_level)
+        raise ValueError(f'Unsupported log level {level_name!r}')
+    _set_log_level(ctx, level, level_name)
 
 @bound_ops("Log")
 def execute_log(ctx: ExecContext, statement: Tree) -> None:
@@ -76,7 +70,6 @@ Log Error "Unexpected value : {!r}", result
 Also see `Print` and `Format()` for formatting details
 
 """
-    from .app_exceptions import VgrRuntimeError
     log_level = statement.children[0].data.title()
     log_func = _LOG_FUNCTION.get(log_level)
     if log_func is None:
@@ -86,3 +79,15 @@ Also see `Print` and `Format()` for formatting details
         format_string = ctx.eval_to_str(statement.children[1], 'Format string', True)
         value = poly_format(format_string, *list(ctx.eval_expr(expr) for expr in statement.children[2:]))
     log_func(value)
+
+def init_app_log(ctx: ExecContext, logname: str) -> None:
+    """Updates the DD with the name and sets the initial log level"""
+    ctx.print_verbose('Log file is', logname)
+    ctx.set_var(logname, *_LOG_FILE_PATH)
+    _set_log_level(ctx, LOG_LEVEL_MAP.get(_DEFAULT_LOG_LEVEL), _DEFAULT_LOG_LEVEL)
+
+def _set_log_level(ctx: ExecContext, level: int, level_name: str) -> None:
+    """Updates the user's logger and the DD"""
+    _USER_LOGGER.setLevel(level)
+    ctx.set_var(level_name, *_LOG_LEVEL_PATH)
+    ctx.print_verbose('Log Level set to', level_name)